backend-django/requirements.txt
# =============================================================================
# Django Backend Requirements
# CRM Application - Python Dependencies
# =============================================================================

# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Caching
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-filter==23.5
argon2-cffi==23.1.0

# API Documentation
drf-spectacular==0.27.0
drf-spectacular[sidecar]==0.27.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
python-decouple==3.8

# Development
django-debug-toolbar==4.2.0
ipython==8.20.0

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0
